Split timed hybrid model and fix Symbolic->Optim dependency (phase 3)
Break the 860-line symbolic_hybrid_model.jl monolith into focused files
under timed_hybrid_symbolic_model/ and move the optimizer-driving builders
out of Symbolic (which must not depend on Optim):

- Symbolic keeps the data structures + pure transition assembly:
  time_symbolic_model.jl, global_input_map.jl,
  timed_hybrid_symbolic_model.jl, hybrid_transition_assembly.jl.
- Optim/HybridSystemAbstraction/hybrid_symbolic_builder.jl now hosts
  build_dynamical_symbolic_model / build_mode_symbolic_abstractions /
  build_timed_hybrid_symbolic_model, removing the lone
  `Dionysos.Optim...` up-reference from Symbolic.

Soundness/cleanup:
- Delete the broad try/catch...@warn;continue blocks that silently dropped
  modes/transitions (build_mode_symbolic_abstractions, add_inter_mode_transitions!,
  find_symbolic_state); legitimate condition-based skips are kept.
- Replace the ~40% commented-out docstrings with real ones.
- Drop leftover debug println.

Hybrid end-to-end suites (time-aware / time-frozen / safety) all green.
