+++ src/symbolic/symbolic.jl
@@ -47,6 +47,12 @@ include("grid_based_symbolic_model/backends/threaded.jl")
 include("grid_based_symbolic_model/backends/julia_distributed.jl")
 include("grid_based_symbolic_model/backends/slurm_array.jl")
 
-include("symbolic_hybrid_model/symbolic_hybrid_model.jl")
+# Timed hybrid symbolic model: data structures + pure transition assembly.
+# (The optimizer-driven builders live in Optim/hybrid_systems, not here, so that
+# Symbolic does not depend on Optim.)
+include("timed_hybrid_symbolic_model/time_symbolic_model.jl")
+include("timed_hybrid_symbolic_model/global_input_map.jl")
+include("timed_hybrid_symbolic_model/timed_hybrid_symbolic_model.jl")
+include("timed_hybrid_symbolic_model/hybrid_transition_assembly.jl")
 
 end  # module Symbolic
